You can be someone's rescue story
The  main motive behind the creation of this application is to save
lives. The application gets information from the Blood-Banks and from
the Organ donor Organizations . The doctors from the hospitals can use
the application to check the inventory of the Blood-Banks for the
availability of blood groups in the vicinity, The same also applies for
the availability of organs or finding a match for a bone marrow
transplant. The application also allows people to register with the
blood group details. The application will alert the volunteer incase of
an emergency , where the hospital is unable to find a match in the
nearby Blood-Banks. The application will also alert the volunteers if a
blood-donation camp or a registration for organ donation is occurring
nearby.